cup_headgear - move each headgear type into separate includes
