+++ addons/cup_headgear/include/6b47/CfgWeapons.hpp
@@ -1,0 +1,227 @@
+// Ratnik Headgear
+class CUP_H_RUS_6B47_v2_Summer: ItemCore
+{
+    class ItemInfo;
+    /*
+    hiddenSelectionsTextures[] = {
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Ra_H_D_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Ra_H_D_Str_co.paa"
+    };
+    */
+};
+class CUP_H_RUS_6B47_v2_GogglesDown_Summer: CUP_H_RUS_6B47_v2_Summer 
+{
+    class ItemInfo;
+    /*
+    hiddenSelectionsTextures[] = {
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Ra_H_S_gd_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Ra_H_S_gd_Str_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+    */
+    
+};
+class CUP_H_RUS_6B47_v2_GogglesUp_Summer: CUP_H_RUS_6B47_v2_Summer 
+{
+    class ItemInfo;
+};
+class CUP_H_RUS_6B47_v2_GogglesClosed_Summer: CUP_H_RUS_6B47_v2_Summer 
+{
+    class ItemInfo;
+};
+
+// 6B47 (Black)
+class tmtm_h_6B47_black: CUP_H_RUS_6B47_v2_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (Black)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_black_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa"
+    };
+};
+class tmtm_h_6B47GogglesDown_black: CUP_H_RUS_6B47_v2_GogglesDown_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (Black/Goggles down)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_black_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesUp_black: CUP_H_RUS_6B47_v2_GogglesUp_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (Black/Goggles up)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_black_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesClosed_black: CUP_H_RUS_6B47_v2_GogglesClosed_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (Black/Goggles closed)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_black_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa"
+    };
+};
+// 6B47 (ATACS AU)
+class tmtm_h_6B47_atacsAu: CUP_H_RUS_6B47_v2_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (ATACS AU)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_atacsAu_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa"
+    };
+};
+class tmtm_h_6B47GogglesDown_atacsAu: CUP_H_RUS_6B47_v2_GogglesDown_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (ATACS AU/Goggles down)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_atacsAu_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesUp_atacsAu: CUP_H_RUS_6B47_v2_GogglesUp_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (ATACS AU/Goggles up)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_atacsAu_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesClosed_atacsAu: CUP_H_RUS_6B47_v2_GogglesClosed_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (ATACS AU/Goggles closed)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_atacsAu_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa"
+    };
+};
+// 6B47 (ATACS FG)
+class tmtm_h_6B47_atacsFg: CUP_H_RUS_6B47_v2_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (ATACS FG)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_atacsFg_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa"
+    };
+};
+class tmtm_h_6B47GogglesDown_atacsFg: CUP_H_RUS_6B47_v2_GogglesDown_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (ATACS FG/Goggles down)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_atacsFg_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesUp_atacsFg: CUP_H_RUS_6B47_v2_GogglesUp_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (ATACS FG/Goggles up)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_atacsFg_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesClosed_atacsFg: CUP_H_RUS_6B47_v2_GogglesClosed_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (ATACS FG/Goggles closed)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_atacsFg_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa"
+    };
+};
+// 6B47 (SURPAT)
+class tmtm_h_6B47_surpat: CUP_H_RUS_6B47_v2_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (SURPAT)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_surpat_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa"
+    };
+};
+class tmtm_h_6B47GogglesDown_surpat: CUP_H_RUS_6B47_v2_GogglesDown_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (SURPAT/Goggles down)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_surpat_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesUp_surpat: CUP_H_RUS_6B47_v2_GogglesUp_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (SURPAT/Goggles up)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_surpat_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesClosed_surpat: CUP_H_RUS_6B47_v2_GogglesClosed_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (SURPAT/Goggles closed)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_surpat_co.paa", 
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmetStrap_black_co.paa"
+    };
+};
+// 6B47 (EMR Winter)
+class tmtm_h_6B47_emrWinter: CUP_H_RUS_6B47_v2_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (EMR Winter)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_emrWinter_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Ra_H_S_Str_co.paa"
+    };
+};
+class tmtm_h_6B47GogglesDown_emrWinter: CUP_H_RUS_6B47_v2_GogglesDown_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (EMR Winter/Goggles down)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_emrWinter_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Ra_H_S_Str_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesUp_emrWinter: CUP_H_RUS_6B47_v2_GogglesUp_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (EMR Winter/Goggles up)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_emrWinter_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Ra_H_S_Str_co.paa", 
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Glass_ca.paa"
+    };
+};
+class tmtm_h_6B47GogglesClosed_emrWinter: CUP_H_RUS_6B47_v2_GogglesClosed_Summer
+{
+    author="Trenchgun";
+    displayName="6B47 (EMR Winter/Goggles closed)";
+    hiddenSelectionsTextures[] = {
+        "\tmtm\pack\cup_headgear\Data\tmtm_ratnik_helmet_emrWinter_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_Russia\RatnikV2\Helmet\Data\Ra_H_S_Str_co.paa"
+    };
+};
