+++ addons/cup_headgear/include/ech/CfgWeapons.hpp
@@ -1,0 +1,706 @@
+// Basic ACH helmet
+class CUP_H_USArmy_HelmetACH_UCP: ItemCore {};
+
+// ACH helmet + headset
+class CUP_H_USArmy_HelmetACH_Headset_UCP: ItemCore {};
+
+// ACH Helmet + Goggles
+class CUP_H_USArmy_HelmetACH_ESS_UCP: ItemCore {};
+
+// ACH Helmet + Headset + Goggles
+class CUP_H_USArmy_HelmetACH_ESS_Headset_UCP: ItemCore {};
+
+// ACH Helmet + Covered Goggles
+class CUP_H_USArmy_HelmetACH_GCOVERED_UCP: ItemCore {};
+
+// ACH Helmet + Covered Goggles + Helmet
+class CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP: ItemCore {};
+
+// MICH 2000 (Black)
+class tmtm_h_ACH_black: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (Black)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_black_co.paa"
+    };
+};
+// MICH 2000 (ATACS AU)
+class tmtm_h_ACH_atacsAu: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS AU)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsAu_co.paa"
+    };
+};
+// MICH 2000 (ATACS FG)
+class tmtm_h_ACH_atacsFg: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS FG)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsFg_co.paa"
+    };
+};
+// MICH 2000 (EMR)
+class tmtm_h_ACH_emr: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emr_co.paa"
+    };
+};
+// MICH 2000 (EMR Winter)
+class tmtm_h_ACH_emrWinter: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR Winter)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emrWinter_co.paa"
+    };
+};
+// MICH 2000 (SURPAT)
+class tmtm_h_ACH_surpat: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (SURPAT)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_surpat_co.paa"
+    };
+};
+// MICH 2000 (White)
+class tmtm_h_ACH_white: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (White)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_whiteGrey_co.paa"
+    };
+};
+
+// MICH 2000 + Headset (Black)
+class tmtm_h_ACHHeadset_black: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (Black/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_black_co.paa"
+    };
+};
+// MICH 2000 + Headset (ATACS AU)
+class tmtm_h_ACHHeadset_atacsAu: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS AU/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsAu_co.paa"
+    };
+};
+// MICH 2000 + Headset (ATACS AU)
+class tmtm_h_ACHHeadset_atacsFg: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS FG/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsFg_co.paa"
+    };
+};
+// MICH 2000 + Headset (EMR)
+class tmtm_h_ACHHeadset_emr: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emr_co.paa"
+    };
+};
+// MICH 2000 + Headset (EMR Winter)
+class tmtm_h_ACHHeadset_emrWinter: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR Winter/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emrWinter_co.paa"
+    };
+};
+// MICH 2000 + Headset (SURPAT)
+class tmtm_h_ACHHeadset_surpat: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (SURPAT/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_surpat_co.paa"
+    };
+};
+// MICH 2000 + Headset (White)
+class tmtm_h_ACHHeadset_white: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (White/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_whiteGrey_co.paa"
+    };
+};
+
+// MICH 2000 + Goggles (Black)
+class tmtm_h_ACHGoggles_black: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (Black/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_black_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles (ATACS AU)
+class tmtm_h_ACHGoggles_atacsAu: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS AU/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsAu_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles (ATACS FG)
+class tmtm_h_ACHGoggles_atacsFg: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS FG/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsFg_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles (EMR)
+class tmtm_h_ACHGoggles_emr: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emr_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles (EMR Winter)
+class tmtm_h_ACHGoggles_emrWinter: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR Winter/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emrWinter_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles (SURPAT)
+class tmtm_h_ACHGoggles_surpat: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (SURPAT/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_surpat_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles (White)
+class tmtm_h_ACHGoggles_white: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (White/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_whiteGrey_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+
+// MICH 2000 + Goggles + Headset (Black)
+class tmtm_h_ACHGogglesHeadset_black: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (Black/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_black_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles + Headset (Black)
+class tmtm_h_ACHGogglesHeadset_atacsAu: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS AU/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsAu_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles + Headset (ATACS FG)
+class tmtm_h_ACHGogglesHeadset_atacsFg: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS FG/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsFg_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles + Headset (EMR)
+class tmtm_h_ACHGogglesHeadset_emr: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emr_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles + Headset (EMR)
+class tmtm_h_ACHGogglesHeadset_emrWinter: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR Winter/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emrWinter_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles + Headset (SURPAT)
+class tmtm_h_ACHGogglesHeadset_surpat: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (SURPAT/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_surpat_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// MICH 2000 + Goggles + Headset (White) 
+class tmtm_h_ACHGogglesHeadset_white: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (White/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_whiteGrey_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+
+// MICH 2000 + Goggle Cover (Black) 
+class tmtm_h_ACHCoveredGoggles_black: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (Black/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_black_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover (ATACS AU) 
+class tmtm_h_ACHCoveredGoggles_atacsAu: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS AU/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsAu_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover (ATACS FG) 
+class tmtm_h_ACHCoveredGoggles_atacsFg: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS FG/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsFg_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover (EMR) 
+class tmtm_h_ACHCoveredGoggles_emr: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emr_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover (EMR Winter)
+class tmtm_h_ACHCoveredGoggles_emrWinter: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR Winter/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emrWinter_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover (SURPAT)
+class tmtm_h_ACHCoveredGoggles_surpat: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (SURPAT/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_surpat_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover (White)
+class tmtm_h_ACHCoveredGoggles_white: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (White/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_whiteGrey_co.paa"
+    };
+};
+
+// MICH 2000 + Goggle Cover + Headset (Black)
+class tmtm_h_ACHCoveredGogglesHeadset_black: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (Black/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_black_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover + Headset (ATACS AU)
+class tmtm_h_ACHCoveredGogglesHeadset_atacsAu: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS AU/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsAu_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover + Headset (ATACS FG)
+class tmtm_h_ACHCoveredGogglesHeadset_atacsFg: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (ATACS FG/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_atacsFg_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover + Headset (EMR)
+class tmtm_h_ACHCoveredGogglesHeadset_emr: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emr_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover + Headset (EMR Winter)
+class tmtm_h_ACHCoveredGogglesHeadset_emrWinter: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (EMR Winter/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_emrWinter_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover + Headset (SURPAT)
+class tmtm_h_ACHCoveredGogglesHeadset_surpat: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (SURPAT/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_surpat_co.paa"
+    };
+};
+// MICH 2000 + Goggle Cover + Headset (White)
+class tmtm_h_ACHCoveredGogglesHeadset_white: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="MICH 2000 (White/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_whiteGrey_co.paa"
+    };
+};
+
+// ECH (MTP)
+class tmtm_h_ECH_mtp: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (MTP)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_mtp_co.paa"
+    };
+};
+// ECH + Headset (MTP)
+class tmtm_h_ECHHeadset_mtp: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (MTP/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_mtp_co.paa"
+    };
+};
+// ECH + Goggles (MTP)
+class tmtm_h_ECHGoggles_mtp: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (MTP/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// ECH + Goggles + Headset (MTP)
+class tmtm_h_ECHGogglesHeadset_mtp: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (MTP/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_mtp_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// ECH + Goggle Cover (MTP) 
+class tmtm_h_ECHCoveredGoggles_mtp: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (MTP/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_mtp_co.paa"
+    };
+};
+// ECH + Goggle Cover + Headset (MTP)
+class tmtm_h_ECHCoveredGogglesHeadset_mtp: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (MTP/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_mtp_co.paa"
+    };
+};
+// ECH (CDF Forest)
+class tmtm_h_ECH_cdfForest: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Forest)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfForest_co.paa"
+    };
+};
+// ECH + Headset (CDF Forest)
+class tmtm_h_ECHHeadset_cdfForest: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Forest/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfForest_co.paa"
+    };
+};
+// ECH + Goggles (CDF Forest)
+class tmtm_h_ECHGoggles_cdfForest: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Forest/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfForest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// ECH + Goggles + Headset (CDF Forest)
+class tmtm_h_ECHGogglesHeadset_cdfForest: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Forest/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfForest_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// ECH + Goggle Cover (CDF Forest) 
+class tmtm_h_ECHCoveredGoggles_cdfForest: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Forest/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfForest_co.paa"
+    };
+};
+// ECH + Goggle Cover + Headset (CDF Forest)
+class tmtm_h_ECHCoveredGogglesHeadset_cdfForest: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Forest/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfForest_co.paa"
+    };
+};
+// ECH (CDF Mountain)
+class tmtm_h_ECH_cdfMountain: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Mountain)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfMountain_co.paa"
+    };
+};
+// ECH + Headset (CDF Mountain)
+class tmtm_h_ECHHeadset_cdfMountain: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Mountain/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfMountain_co.paa"
+    };
+};
+// ECH + Goggles (CDF Mountain)
+class tmtm_h_ECHGoggles_cdfMountain: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Mountain/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfMountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// ECH + Goggles + Headset (CDF Mountain)
+class tmtm_h_ECHGogglesHeadset_cdfMountain: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Mountain/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfMountain_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// ECH + Goggle Cover (CDF Mountain) 
+class tmtm_h_ECHCoveredGoggles_cdfMountain: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Mountain/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfMountain_co.paa"
+    };
+};
+// ECH + Goggle Cover + Headset (CDF Mountain)
+class tmtm_h_ECHCoveredGogglesHeadset_cdfMountain: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Mountain/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfMountain_co.paa"
+    };
+};
+// ECH (CDF Oxblood)
+class tmtm_h_ECH_cdfOxblood: CUP_H_USArmy_HelmetACH_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Oxblood)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfOxblood_co.paa"
+    };
+};
+// ECH + Headset (CDF Oxblood)
+class tmtm_h_ECHHeadset_cdfOxblood: CUP_H_USArmy_HelmetACH_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Oxblood/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfOxblood_co.paa"
+    };
+};
+// ECH + Goggles (CDF Oxblood)
+class tmtm_h_ECHGoggles_cdfOxblood: CUP_H_USArmy_HelmetACH_ESS_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Oxblood/Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfOxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// ECH + Goggles + Headset (CDF Oxblood)
+class tmtm_h_ECHGogglesHeadset_cdfOxblood: CUP_H_USArmy_HelmetACH_ESS_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Oxblood/Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfOxblood_co.paa",
+        "\CUP\Creatures\People\Military\CUP_Creatures_People_Military_USArmy\data\h_gear_01_rgr_co.paa"
+    };
+};
+// ECH + Goggle Cover (CDF Oxblood) 
+class tmtm_h_ECHCoveredGoggles_cdfOxblood: CUP_H_USArmy_HelmetACH_GCOVERED_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Oxblood/Covered Goggles)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfOxblood_co.paa"
+    };
+};
+// ECH + Goggle Cover + Headset (CDF Oxblood)
+class tmtm_h_ECHCoveredGogglesHeadset_cdfOxblood: CUP_H_USArmy_HelmetACH_GCOVERED_Headset_UCP
+{
+    author="Trenchgun";
+    scope=2;
+    displayName="ECH (CDF Oxblood/Covered Goggles/Headset)";
+    hiddenSelectionsTextures[]={
+        "\tmtm\pack\cup_headgear\Data\tmtm_mich2000_cdfOxblood_co.paa"
+    };
+};
